+++ admin/language/english/payment/snapio.php
@@ -19,9 +19,6 @@ $_['entry_server_key']   = 'Server Key'; // v2 API only
 $_['entry_client_key']   = 'Client Key'; // v2 API only
 $_['entry_3d_secure']    = 'Enable 3D Secure?';
 $_['entry_payment_type'] = 'Payment Type';
-$_['entry_test']         = 'Test Mode:';
-$_['entry_total']        = 'Total:<br /><span class="help">The checkout total the order must reach before this payment method becomes active.</span>';
-$_['entry_order_status'] = 'Order Status:';
 $_['entry_geo_zone']     = 'Geo Zone';
 $_['entry_status']       = 'Status';
 $_['entry_sort_order']   = 'Sort Order';
@@ -32,6 +29,7 @@ $_['entry_acq_bank']	 = 'Acquiring Bank';
 $_['entry_installment_term'] = 'Installment Terms';
 $_['entry_bin_number']  = 'Bin Number';
 $_['entry_custom_field'] = 'Custom Field';
+$_['entry_mixpanel']	 = 'Midtrans Mixpanel';
 
 // Help
 $_['help_min'] = 'Minimum amount of transaction.';